refactor(simulation): decompose EncounterHistory into focused sub-components
- Created encounter-history components:
  - BattleListSidebar.tsx (259L) - battle list and selection
  - ComparisonSection.tsx (234L) - force comparison display
  - EventTimelineSection.tsx (191L) - event timeline with VCR controls
  - DamageMatrixSection.tsx (112L) - damage matrix visualization
  - ForcesSection.tsx (108L) - forces display
  - KeyMomentsSection.tsx (95L) - key moments display
- Reduced EncounterHistory.tsx from 1,248L to 146L (88% reduction)
- All 128 tests passing
- Largest component: 259L (< 400L limit)
