+++ src/components/simulation-viewer/pages/encounter-history/index.ts
@@ -1,0 +1,20 @@
+export { BattleListSidebar } from './BattleListSidebar';
+export { ForcesSection } from './ForcesSection';
+export { DamageMatrixSection } from './DamageMatrixSection';
+export { KeyMomentsSection } from './KeyMomentsSection';
+export { EventTimelineSection } from './EventTimelineSection';
+export { ComparisonSection } from './ComparisonSection';
+
+export type {
+  IEncounterHistoryProps,
+  IBattle,
+  IForce,
+  IDamageMatrix,
+  IKeyMoment,
+  IBattleEvent,
+  SortKey,
+  SortDirection,
+  ComparisonMode,
+} from './types';
+
+export { formatDuration } from './types';
